project/requirements.txt
beautifulsoup4==4.9.3
boto3==1.18.1
Django==3.1.9
django-rest-auth==0.9.5
django-allauth==0.45.0
django-cors-headers==3.7.0
dj_rest_auth==2.1.9
django-mysql==3.12.0
django-storages==1.11.1
django-grappelli==2.15.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pillow==8.3.1
dj-rest-auth==2.1.9
gunicorn==20.1.0
uvicorn==0.14.0
mysqlclient==2.0.3
redis==3.5.3
whitenoise==5.3.0